working joycons

+++ src/usb_descriptors.h
@@ -4,7 +4,7 @@
 // Custom HID struct matching our descriptor
 // Renamed to avoid collision with TinyUSB's default hid_gamepad_report_t
 typedef struct __attribute__((packed)) {
-    uint16_t buttons; // 16 Buttons
+    uint32_t buttons; // 32 Buttons (Expanded from 16 to match standard PC layouts)
     uint8_t  hat;     // 8-way D-pad (0 = centered)
     uint16_t x;       // Left Stick X (0-4095)
     uint16_t y;       // Left Stick Y (0-4095)
